Add common mock post file
