--- packages/grid/components/GridPosts/postData.js
@@ -1,16 +1,0 @@
-export const gridPosts = [
-  {
-    id: '5c829d3e38da0900fe311142',
-    thumbnail: 'https://buffer-media-uploads-dev.s3.amazonaws.com/5b8e886dbee2c512007b23c6/5c829d3838da0900d16ee5e4/3a9dd6c260165524ba20b2fd174a0873.original.jpg',
-    sent_at: 1552569720,
-    scheduled: false,
-    link: null,
-  },
-  {
-    id: '5c8284f138da0900470d81a2',
-    thumbnail: 'https://buffer-media-uploads-dev.s3.amazonaws.com/5b8e886dbee2c512007b23c6/5c827fa938da09002100e972/7d8cce873bd0da0d10a4aeb683f99149.original.jpg',
-    due_at: 1553276340,
-    scheduled: true,
-    link: 'https://www.page.com',
-  },
-];
