+++ packages/past-reminders/components/PastRemindersPosts/postData.js
@@ -1,163 +1,3 @@
-export const postLists = [
-  {
-    queueItemType: 'header',
-    text: 'Today',
-    id: 'header-1',
-  },
-  {
-    queueItemType: 'post',
-    id: '590a365d749c2018007b23c6',
-    isConfirmingDelete: false,
-    isDeleting: false,
-    isWorking: false,
-    imageUrls: [],
-    isPastReminder: true,
-    links: [],
-    linkAttachment: {},
-    postContent: {
-      imageUrls: [],
-      linkAttachment: {},
-      links: [],
-      retweetCommentLinks: [],
-      text: 'New thing',
-      type: 'text',
-    },
-    postDetails: {
-      postAction: 'This post is scheduled for May 2nd',
-      isRetweet: false,
-    },
-    retweetCommentLinks: [],
-    sent: false,
-    text: 'New thing',
-    type: 'text',
-    day: 'Tuesday 12th February',
-    dueTime: '9:02 am',
-  },
-  {
-    queueItemType: 'post',
-    id: '590a3693749c200e007b23c7',
-    isConfirmingDelete: false,
-    isDeleting: false,
-    isWorking: false,
-    imageUrls: [],
-    links: [],
-    linkAttachment: {},
-    postContent: {
-      imageUrls: [],
-      linkAttachment: {},
-      links: [],
-      retweetCommentLinks: [],
-      text: 'Another thing, that is also new',
-      type: 'text',
-    },
-    postDetails: {
-      postAction: 'This post is scheduled for May 2nd',
-      isRetweet: false,
-    },
-    retweetCommentLinks: [],
-    sent: true,
-    text: 'Another thing, that is also new',
-    type: 'text',
-    day: 'Tuesday 12th February',
-    dueTime: '9:02 am',
-  },
-  {
-    queueItemType: 'header',
-    text: 'Tuesday May 3rd',
-    id: 'header-2',
-  },
-  {
-    queueItemType: 'post',
-    id: '590a365d749c2018007b23c8',
-    isConfirmingDelete: false,
-    isDeleting: false,
-    isWorking: false,
-    imageUrls: [],
-    links: [],
-    linkAttachment: {},
-    postContent: {
-      imageSrc:
-        'https://cdn-images-1.medium.com/max/2000/1*1Kua7bNJfvLlTxWqgxVKfw.jpeg',
-      imageUrls: [
-        'https://cdn-images-1.medium.com/max/2000/1*1Kua7bNJfvLlTxWqgxVKfw.jpeg',
-      ],
-      linkAttachment: {},
-      links: [],
-      retweetCommentLinks: [],
-      text: 'Another thing, that is also new',
-      type: 'image',
-    },
-    postDetails: {
-      postAction: 'This post is scheduled for May 3rd',
-      isRetweet: false,
-    },
-    retweetCommentLinks: [],
-    sent: true,
-    text: 'Image post test',
-    type: 'image',
-    imageSrc:
-      'https://cdn-images-1.medium.com/max/2000/1*1Kua7bNJfvLlTxWqgxVKfw.jpeg',
-    day: 'Tuesday 12th February',
-    dueTime: '9:02 am',
-  },
-  {
-    queueItemType: 'post',
-    id: '590a3693749c200e007b23c9',
-    isConfirmingDelete: false,
-    isDeleting: false,
-    isWorking: false,
-    imageUrls: [],
-    links: [],
-    linkAttachment: {},
-    postContent: {
-      imageUrls: [],
-      linkAttachment: {},
-      links: [],
-      retweetCommentLinks: [],
-      text: 'Another thing, that is also new',
-      type: 'text',
-    },
-    postDetails: {
-      postAction: 'This post is scheduled for May 3rd',
-      isRetweet: false,
-    },
-    retweetCommentLinks: [],
-    sent: true,
-    text: 'Another thing, that is also new',
-    type: 'text',
-    day: 'Tuesday 12th February',
-    dueTime: '9:02 am',
-  },
-  {
-    queueItemType: 'post',
-    id: '590a3693749c200e007b23c10',
-    isConfirmingDelete: false,
-    isDeleting: false,
-    isWorking: false,
-    imageUrls: [],
-    links: [],
-    linkAttachment: {},
-    postContent: {
-      imageUrls: [],
-      linkAttachment: {},
-      links: [],
-      retweetCommentLinks: [],
-      text: 'Another thing, that is also new',
-      type: 'text',
-    },
-    postDetails: {
-      postAction: 'This post is scheduled for May 4th',
-      isRetweet: false,
-    },
-    retweetCommentLinks: [],
-    sent: true,
-    text: 'Another thing, that is also new',
-    type: 'text',
-    day: 'Tuesday 12th February',
-    dueTime: '9:02 am',
-  },
-];
-
 export const header = 'Recent past reminders';
 export const subHeader =
   'Here you can manage posts that may still require posting from your mobile';
