--- packages/queue/components/QueuedPosts/postData.js
@@ -1,122 +1,0 @@
-const items = [
-  {
-    queueItemType: 'header',
-    text: 'Today',
-    id: 'header-1',
-  },
-  {
-    queueItemType: 'post',
-    id: '590a365d749c2018007b23c6',
-    isConfirmingDelete: false,
-    isDeleting: false,
-    isWorking: false,
-    imageUrls: [],
-    links: [],
-    linkAttachment: {},
-    postContent: {
-      imageUrls: [],
-      linkAttachment: {},
-      links: [],
-      retweetCommentLinks: [],
-      text: 'New thing',
-      type: 'text',
-    },
-    postDetails: {
-      postAction: 'This post is scheduled for May 2nd',
-      isRetweet: false,
-      error: '',
-    },
-    retweetCommentLinks: [],
-    sent: false,
-    text: 'New thing',
-    type: 'text',
-  },
-  {
-    queueItemType: 'post',
-    id: '590a3693749c200e007b23c7',
-    isConfirmingDelete: false,
-    isDeleting: false,
-    isWorking: false,
-    imageUrls: [],
-    links: [],
-    linkAttachment: {},
-    postContent: {
-      imageUrls: [],
-      linkAttachment: {},
-      links: [],
-      retweetCommentLinks: [],
-      text: 'Another thing, that is also new',
-      type: 'text',
-    },
-    postDetails: {
-      postAction: 'This post is scheduled for May 2nd',
-      isRetweet: false,
-      error: '',
-    },
-    retweetCommentLinks: [],
-    sent: false,
-    text: 'Another thing, that is also new',
-    type: 'text',
-  },
-  {
-    queueItemType: 'header',
-    text: 'Tuesday May 3rd',
-    id: 'header-2',
-  },
-  {
-    queueItemType: 'post',
-    id: '590a365d749c2018007b23c1',
-    isConfirmingDelete: false,
-    isDeleting: false,
-    isWorking: false,
-    imageUrls: [],
-    links: [],
-    linkAttachment: {},
-    postContent: {
-      imageUrls: [],
-      linkAttachment: {},
-      links: [],
-      retweetCommentLinks: [],
-      text: 'New thing',
-      type: 'text',
-    },
-    postDetails: {
-      postAction: 'This post is scheduled for May 3rd',
-      isRetweet: false,
-      error: '',
-    },
-    retweetCommentLinks: [],
-    sent: false,
-    text: 'New thing',
-    type: 'text',
-  },
-  {
-    queueItemType: 'post',
-    id: '590a3693749c200e007b23c2',
-    isConfirmingDelete: false,
-    isDeleting: false,
-    isWorking: false,
-    imageUrls: [],
-    links: [],
-    linkAttachment: {},
-    postContent: {
-      imageUrls: [],
-      linkAttachment: {},
-      links: [],
-      retweetCommentLinks: [],
-      text: 'Another thing, that is also new',
-      type: 'text',
-    },
-    postDetails: {
-      postAction: 'This post is scheduled for May 3rd',
-      isRetweet: false,
-      error: '',
-    },
-    retweetCommentLinks: [],
-    sent: false,
-    text: 'Another thing, that is also new',
-    type: 'text',
-  },
-];
-
-export default items;
