--- packages/sent/components/SentPosts/postData.js
@@ -1,146 +1,0 @@
-const postLists = [
-  {
-    queueItemType: 'header',
-    text: 'Today',
-    id: 'header-1',
-  },
-  {
-    queueItemType: 'post',
-    id: '590a365d749c2018007b23c6',
-    isConfirmingDelete: false,
-    isDeleting: false,
-    isWorking: false,
-    imageUrls: [],
-    links: [],
-    linkAttachment: {},
-    postContent: {
-      imageUrls: [],
-      linkAttachment: {},
-      links: [],
-      retweetCommentLinks: [],
-      text: 'New thing',
-      type: 'text',
-    },
-    postDetails: {
-      postAction: 'This post was sent May 2nd',
-      isRetweet: false,
-    },
-    retweetCommentLinks: [],
-    sent: true,
-    text: 'New thing',
-    type: 'text',
-    statistics: {
-      retweets: 10,
-      favorites: 29,
-      mentions: 0,
-      clicks: 1,
-      reach: 3,
-    },
-  },
-  {
-    queueItemType: 'post',
-    id: '590a3693749c200e007b23c7',
-    isConfirmingDelete: false,
-    isDeleting: false,
-    isWorking: false,
-    imageUrls: [],
-    links: [],
-    linkAttachment: {},
-    postContent: {
-      imageUrls: [],
-      linkAttachment: {},
-      links: [],
-      retweetCommentLinks: [],
-      text: 'Another thing, that is also new',
-      type: 'text',
-    },
-    postDetails: {
-      postAction: 'This post was sent May 2nd',
-      isRetweet: false,
-    },
-    retweetCommentLinks: [],
-    sent: true,
-    text: 'Another thing, that is also new',
-    type: 'text',
-    statistics: {
-      retweets: 10,
-      favorites: 29,
-      mentions: 0,
-      clicks: 1,
-      reach: 3,
-    },
-  },
-  {
-    queueItemType: 'header',
-    text: 'Yesterday, May 3',
-    id: 'header-2',
-  },
-  {
-    queueItemType: 'post',
-    id: '590a365d749c2018007b23c8',
-    isConfirmingDelete: false,
-    isDeleting: false,
-    isWorking: false,
-    imageUrls: [],
-    links: [],
-    linkAttachment: {},
-    postContent: {
-      imageUrls: [],
-      linkAttachment: {},
-      links: [],
-      retweetCommentLinks: [],
-      text: 'New thing',
-      type: 'text',
-    },
-    postDetails: {
-      postAction: 'This post was sent May 3rd',
-      isRetweet: false,
-    },
-    retweetCommentLinks: [],
-    sent: true,
-    text: 'New thing',
-    type: 'text',
-    statistics: {
-      retweets: 10,
-      favorites: 29,
-      mentions: 0,
-      clicks: 1,
-      reach: 3,
-    },
-  },
-  {
-    queueItemType: 'post',
-    id: '590a3693749c200e007b239',
-    isConfirmingDelete: false,
-    isDeleting: false,
-    isWorking: false,
-    imageUrls: [],
-    links: [],
-    linkAttachment: {},
-    postContent: {
-      imageUrls: [],
-      linkAttachment: {},
-      links: [],
-      retweetCommentLinks: [],
-      text: 'Another thing, that is also new',
-      type: 'text',
-    },
-    postDetails: {
-      postAction: 'This post was sent May 3rd',
-      isRetweet: false,
-    },
-    retweetCommentLinks: [],
-    sent: true,
-    text: 'Another thing, that is also new',
-    type: 'text',
-    statistics: {
-      retweets: 10,
-      favorites: 29,
-      mentions: 0,
-      clicks: 1,
-      reach: 3,
-    },
-  },
-];
-
-export default postLists;
