--- packages/stories/components/StoryGroups/storiesData.js
@@ -1,117 +1,0 @@
-const storyGroups = [
-  {
-    date: 'September 16',
-    dayOfWeek: 'Today',
-    id: 'Today',
-    queueItemType: 'header',
-    text: 'Today',
-  },
-  {
-    queueItemType: 'post',
-    _id: '5d836482869bb3001e64c0c3',
-    createdAt: 1568892034,
-    day: 'Today',
-    due_time: '4:20 pm',
-    id: '5d836482869bb3001e64c0c3',
-    profile_id: '5c81476338da0901743a63a2',
-    profile_timezone: 'Europe/Madrid',
-    queue_nonce: '1568892034',
-    scheduled_at: 1568902800,
-    status: 'scheduled',
-    storyDetails: {
-      avatarUrl:
-        'https://secure.gravatar.com/avatar/03cd143dc214b358e032de8a136a6027?s=40&d=mm',
-      createdAt: 'today at 10:57 AM (BST)',
-      creatorName: 'mayita.uribe',
-      dueTime: '10:00 am',
-      status: 'scheduled',
-      storyAction:
-        "You will receive a reminder September 17th at 10:00 AM (BST) when it's time to post.",
-      stories: [
-        {
-          _id: '5d836482869bb3001e64c0c2',
-          id: '5d836482869bb3001e64c0c2',
-          note: '',
-          type: 'image',
-          order: 1,
-          asset_url:
-            'https://buffer-media-uploads-dev.s3.amazonaws.com/5b8e886dbee2c512007b23c6/5d83647c869bb300116fc12a/d14d93b05119af69fd9d68fafe839157.original.jpg',
-          thumbnail_url:
-            'https://buffer-media-uploads-dev.s3.amazonaws.com/5b8e886dbee2c512007b23c6/5d83647c869bb300116fc12a/d14d93b05119af69fd9d68fafe839157.original.jpg',
-        },
-      ],
-    },
-    type: 'storyGroup',
-    twentyfour_hour_time: false,
-    updated_at: 1568892034,
-    user_avatar:
-      'https://secure.gravatar.com/avatar/03cd143dc214b358e032de8a136a6027?s=40&d=mm',
-    user_id: '5b8e886dbee2c512007b23c6',
-    user_name: 'mayita.uribe',
-  },
-  {
-    queueItemType: 'post',
-    _id: '5d725b3e3bd46f65105ae76b',
-    created_at: 1567584428,
-    day: 'Tomorrow',
-    due_time: '7:45 pm',
-    id: '5d725b3e3bd46f65105ae76b',
-    profile_id: '5c81476338da0901743a63a2',
-    profile_timezone: 'Europe/Madrid',
-    scheduled_at: 1569001500,
-    status: 'scheduled',
-    storyDetails: {
-      avatarUrl:
-        'https://secure.gravatar.com/avatar/03cd143dc214b358e032de8a136a6027?s=40&d=mm',
-      createdAt: 'today at 10:57 AM (BST)',
-      creatorName: 'mayita.uribe',
-      dueTime: '10:00 am',
-      status: 'scheduled',
-      storyAction:
-        "You will receive a reminder September 17th at 10:00 AM (BST) when it's time to post.",
-      stories: [
-        {
-          _id: '5d8365258943cc4bc52d94cd',
-          id: '5d8365258943cc4bc52d94cd',
-          note: 'Puchi happy',
-          type: 'image',
-          order: 1,
-          asset_url:
-            'https://buffer-media-uploads-dev.s3.amazonaws.com/5b8e886dbee2c512007b23c6/5d5f9477869bb3305737a3f2/dd3a25447a6e5c305ae40e08388e5bc2.original.jpg',
-          thumbnail_url:
-            'https://buffer-media-uploads-dev.s3.amazonaws.com/5b8e886dbee2c512007b23c6/5d5f9477869bb3305737a3f2/dd3a25447a6e5c305ae40e08388e5bc2.original.jpg',
-        },
-        {
-          _id: '5d83653d8943cc4bc52d94ce',
-          id: '5d83653d8943cc4bc52d94ce',
-          note: 'Note',
-          type: 'image',
-          order: 2,
-          asset_url:
-            'https://buffer-media-uploads-dev.s3.amazonaws.com/5b8e886dbee2c512007b23c6/5d5f9c64869bb331a32f5552/7f6dd5f725507d760a782527adacb055.original.jpg',
-          thumbnail_url:
-            'https://buffer-media-uploads-dev.s3.amazonaws.com/5b8e886dbee2c512007b23c6/5d5f9c64869bb331a32f5552/7f6dd5f725507d760a782527adacb055.original.jpg',
-        },
-        {
-          _id: '5d8365528943cc4bc52d94cf',
-          id: '5d8365528943cc4bc52d94cf',
-          note: 'Puchi note',
-          type: 'image',
-          order: 3,
-          asset_url:
-            'https://buffer-media-uploads-dev.s3.amazonaws.com/5b8e886dbee2c512007b23c6/5d4982176826f315ed1e79ba/f458295a6daad1c702af20c0b013defd.original.jpg',
-          thumbnail_url:
-            'https://buffer-media-uploads-dev.s3.amazonaws.com/5b8e886dbee2c512007b23c6/5d4982176826f315ed1e79ba/f458295a6daad1c702af20c0b013defd.original.jpg',
-        },
-      ],
-    },
-    twentyfour_hour_time: false,
-    user_avatar:
-      'https://secure.gravatar.com/avatar/03cd143dc214b358e032de8a136a6027?s=40&d=mm',
-    user_name: 'mayita.uribe',
-    user_id: '5b8e886dbee2c512007b23c6',
-    type: 'storyGroup',
-  },
-];
-
-export default storyGroups;
